move files around with other transformers, and make Symbol into a const.
