+++ packages/jsii/lib/assembler.ts
@@ -20,8 +20,8 @@ import * as literate from './literate';
 import * as bindings from './node-bindings';
 import { ProjectInfo } from './project-info';
 import { isReservedName } from './reserved-words';
-import { RuntimeTypeInfoInjector } from './runtime-info';
 import { DeprecatedRemover } from './transforms/deprecated-remover';
+import { RuntimeTypeInfoInjector } from './transforms/runtime-info';
 import { TsCommentReplacer } from './transforms/ts-comment-replacer';
 import { combinedTransformers } from './transforms/utils';
 import { Validator } from './validator';
